Revert "Revert ABI2, it wasn't meant to end up in master branch."
This reverts commit 950d210ec5d1d49940f60600496e43b9db510f5e.

+++ common/const.h
@@ -14,6 +14,9 @@
 ****/
 #ifndef CONST_H
 #define CONST_H
+
+#include <stddef.h> // ptrdiff_t
+
 //
 // Constants shared by the engine and dlls
 // This header file included by engine files and DLL files.
@@ -721,7 +724,7 @@ enum
 };
 
 typedef int		func_t;
-typedef int		string_t;
+typedef ptrdiff_t	string_t;
 
 typedef unsigned short	word;
